Add new C# project "Malshinon" with configuration files
This commit introduces a new C# project named "Malshinon."
It includes an `App.config` for runtime settings, a project file
(`Malshinon.csproj`) specifying properties and references, and
a solution file (`Malshinon.sln`) for organization in Visual Studio.
Additionally, a basic `Program.cs` file with a minimal `Main`
method and an `AssemblyInfo.cs` file for assembly metadata
are included.